Bump to anchor 0.24.2 (#69)
* Bump to anchor 0.24.2

Bump from Anchor 0.18.2 to 0.24.2

Anchor introduces a number of breaking changes since 18, primarily,
the Provider class has been replaced with an AnchorProvider class which
implements Provider as an interface. This PR contains the necessary
changes to handle this upgrade.

Exported enums in each directory have been replaced with AnchorTypes
from the saberhq/anchor-contrib directory. This is done to match the
type expected when calling `program.accounts.SomeAccount.fetch()`

* Remove atlas-staking import

* Npm audit fix